Assistant checkpoint: Update valuation calculation to use reactive variable
Assistant generated file changes:
- src/views/ValuationDetail.vue: Update doValuationCalculation to use reactive variable

---

User prompt:

in the function doValuationCalculation i do not want to return the values for each calculated metric but rather i want to save them in the reactiveVariable valuationCalculation.

e.g. in valuationCalculation.futureGrowthRate, valuationCalculation.total_arr_multiple_impact etc.

please amend the code so that upon opening the page all of the values within doValuationCalculation are being calculated and stored in the reactiveVariable valuationCalculation.