Assistant checkpoint: Move URL formatting to utils and update components
Assistant generated file changes:
- src/utils/url_formatting.js: Create new URL formatting utilities
- src/views/ValuationDetail.vue: Update URL handling to use imported functions
- src/views/Valuations.vue: Add URL formatting to Valuations view, Update script section with URL formatting, Add URL formatting function

---

User prompt:

in ValuationsDetail.vue the user can input a company url.

i have one or multiple helper functions that helps the user to save a valid url and to ouput a valid url in the template.

please move those helper to a new file under /utils/ called "url_formatting.js".

import those functions to ValuationDetail.vue so that the site works as it is at the moment.

lastly import the same functions to Valuations.vue and apply them to improve the Company URL input field as well as the Company URL Output in the template.