Assistant checkpoint: Add financial info navigation and conditional visibility
Assistant generated file changes:
- src/views/ValuationDetail.vue: Add navigateToFinancialInfo function and update template conditions, Add navigateToFinancialInfo function

---

User prompt:

in @src/views/ValuationDetail.vue enhance the functionality of the v-btn that includes @click="navigateToFinancialInfo" as follows.

upon click of the button the function navigateToFinancialInfo should be called.

create that function.

when the user clicks the button the reactive variable valuationData.show_financial_inputs should be set to true.

the v-if statement in the template that is currently "<template
      v-if="valuationData.valuation_type !== null && valuationData.valuation_type !== ''"
      >" should be changed so that the statement returns true, if valuationData.show_financial_inputs is true

i.e. only render the content of the template if   valuationData.show_financial_inputs is true.

lastly dont show the button that includes @click="navigateToFinancialInfo" after the button has been pressed.