Assistant checkpoint: Add inline URL editing functionality
Assistant generated file changes:
- src/views/ValuationDetail.vue: Add inline editing for company URL, Add URL editing state variables and methods

---

User prompt:

change the functionality of @src/views/ValuationDetail.vue as follows.

currently the template element

<a :href="valuationData.company_url" target="_blank" class="text-body-1 mr-6">{{ valuationData.company_url }}</a>

only displays the company_url.

add a little mdi edit-icon next to the company_url.

when the user click that edit icon, the company_url link changes to an input field that has the company_url as a value.

apply all rules that are applied in

        <v-text-field
          v-model="valuationData.company_url"
          id="company_url"
          label="Company Website"
          required
          :rules="[urlRules]"
          placeholder="www.example.com"
          @blur="formatURL"
          hide-details="auto"
          class="mb-6"
        ></v-text-field>

        to the new input field.

as the user changes the content of valuationData.company_url the current watch should detect that change and save the data.

close the input field when the user clicks outside of the input field.